Move CompiledNode
